Fix interrupt detecting issues

+++ Firmwares/HA4IoT.Controller.WiFiOutpost/src/Mqtt.h
@@ -1,7 +1,7 @@
 #include <Arduino.h>
 
 void setupMqtt();
-void loopMqtt(unsigned int elapsedMillis);
+void loopMqtt(uint16_t elapsedMillis);
 
 bool getMqttIsConnected();
